requirements.txt
ipykernel
langchain
openai
faiss-cpu
tiktoken
azure-devops
python-dotenv
beautifulsoup4
